Issue #3028367 by chmez, kingdutch: move to a single configuration override class
The work done for the social_embed configuration override is pretty
important to happen together. It makes sense to just move this work into
a single class with separate methods that are easy to parse. This
reduces the cognitive overhead of having a base class and further
indirection.
