+++ modules/social_features/social_embed/social_embed.services.yml
@@ -4,8 +4,3 @@ services:
     arguments: ['@module_handler', '@config.factory']
     tags:
       - {name: config.factory.override, priority: 50}
-  social_embed_editor.overrider:
-    class: \Drupal\social_embed\SocialEmbedEditorConfigOverride
-    arguments: ['@module_handler', '@config.factory']
-    tags:
-      - {name: config.factory.override, priority: 51}
