+++ modules/social_features/social_embed/src/SocialEmbedConfigOverride.php
@@ -2,19 +2,87 @@
 
 namespace Drupal\social_embed;
 
-use Drupal\Core\Config\Config;
+use Drupal\Core\Cache\CacheableMetadata;
+use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Config\ConfigFactoryOverrideInterface;
+use Drupal\Core\Config\StorageInterface;
+use Drupal\Core\Extension\ModuleHandlerInterface;
 
 /**
  * Class SocialEmbedConfigOverride.
  *
  * @package Drupal\social_embed
  */
-class SocialEmbedConfigOverride extends SocialEmbedConfigOverrideBase {
+class SocialEmbedConfigOverride implements ConfigFactoryOverrideInterface {
+
+  /**
+   * The module handler.
+   *
+   * @var \Drupal\Core\Extension\ModuleHandlerInterface
+   */
+  protected $moduleHandler;
+
+  /**
+   * The config factory.
+   *
+   * @var \Drupal\Core\Config\ConfigFactoryInterface
+   */
+  protected $configFactory;
+
+  /**
+   * Constructs the configuration override.
+   *
+   * @param \Drupal\Core\Extension\ModuleHandlerInterface $module_handler
+   *   The module handler.
+   * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
+   *   The configuration factory.
+   */
+  public function __construct(ModuleHandlerInterface $module_handler, ConfigFactoryInterface $config_factory) {
+    $this->moduleHandler = $module_handler;
+    $this->configFactory = $config_factory;
+  }
 
   /**
    * {@inheritdoc}
    */
-  public function doOverride(Config $config, $config_name, $convert_url, array &$overrides) {
+  public function loadOverrides($names) {
+    $overrides = [];
+
+    $formats = [
+      'basic_html' => TRUE,
+      'full_html' => FALSE,
+    ];
+
+    $this->moduleHandler->alter('social_embed_formats', $formats);
+
+    foreach ($formats as $format => $convert_url) {
+      if (in_array('filter.format.' . $format, $names)) {
+        $this->addFilterOverride($format, $convert_url, $overrides);
+      }
+
+      if (in_array('editor.editor.' . $format, $names)) {
+        $this->addEditorOverride($format, $overrides);
+      }
+    }
+
+    return $overrides;
+  }
+
+  /**
+   * Alters the filter settings for the text format.
+   *
+   * @param string $text_format
+   *   A config name.
+   * @param bool $convert_url
+   *   TRUE if filter should be used.
+   * @param array $overrides
+   *   An override configuration.
+   */
+  protected function addFilterOverride($text_format, $convert_url, array &$overrides) {
+    $config_name = 'filter.format.' . $text_format;
+    /* @var \Drupal\Core\Config\Config $config */
+    $config = $this->configFactory->getEditable($config_name);
+
     $dependencies = $config->get('dependencies.module');
     $dependencies[] = 'url_embed';
 
@@ -51,17 +119,71 @@ class SocialEmbedConfigOverride extends SocialEmbedConfigOverrideBase {
   }
 
   /**
-   * {@inheritdoc}
+   * Alters the editor settings for the text format.
+   *
+   * @param string $text_format
+   *   The text format to adjust.
+   * @param array $overrides
+   *   An override configuration.
    */
-  public function getCacheSuffix() {
-    return 'SocialEmbedConfigOverride';
+  protected function addEditorOverride($text_format, array &$overrides) {
+    $config_name = 'editor.editor.' . $text_format;
+    /* @var \Drupal\Core\Config\Config $config */
+    $config = $this->configFactory->getEditable($config_name);
+    $settings = $config->get('settings');
+
+    if (!($settings && isset($settings['toolbar']['rows']) && is_array($settings['toolbar']['rows']))) {
+      return;
+    }
+
+    $button_exists = FALSE;
+
+    foreach ($settings['toolbar']['rows'] as $row) {
+      foreach ($row as $group) {
+        foreach ($group['items'] as $button) {
+          if ($button === 'social_embed') {
+            $button_exists = TRUE;
+            break;
+          }
+        }
+      }
+    }
+
+    if (!$button_exists) {
+      $row_array_keys = array_keys($settings['toolbar']['rows']);
+      $last_row_key = end($row_array_keys);
+
+      $group = [];
+      $group['name'] = 'Embed';
+      $group['items'] = [];
+      $group['items'][] = 'social_embed';
+      $settings['toolbar']['rows'][$last_row_key][] = $group;
+
+      $overrides[$config_name] = [
+        'settings' => $settings,
+      ];
+    }
   }
 
   /**
    * {@inheritdoc}
    */
-  public function getPrefix() {
-    return 'filter.format';
+  public function getCacheableMetadata($name) {
+    return new CacheableMetadata();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function createConfigObject($name, $collection = StorageInterface::DEFAULT_COLLECTION) {
+    return NULL;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getCacheSuffix() {
+    return 'SocialEmbedConfigOverride';
   }
 
 }
